Move files around and cleanup structure
