--- are_we_proc_macro
@@ -1,2 +1,0 @@
-Hell Yeah - TrueDoctor, 2021
-Hell Yeah - T0mstone, 2021
