+++ client/web/wasm/src/window.rs
@@ -20,12 +20,12 @@ pub fn get_active_document() -> DocumentId {
 	todo!("get_active_document")
 }
 
-use editor_core::workspace::PanelId;
+/*use editor_core::workspace::PanelId;
 /// Notify the editor that the mouse hovers above a panel
 #[wasm_bindgen]
 pub fn panel_hover_enter(panel_id: PanelId) {
 	todo!("panel_hover_enter {}", panel_id)
-}
+}*/
 
 /// Query a list of currently available operations
 #[wasm_bindgen]
