+++ core/editor/src/frontend/mod.rs
@@ -1,0 +1,3 @@
+pub mod frontend_message_handler;
+
+pub use frontend_message_handler::{Callback, FrontendMessage, FrontendMessageDiscriminant, FrontendMessageHandler};
