+++ core/editor/src/global/mod.rs
@@ -1,0 +1,3 @@
+pub mod global_message_handler;
+
+pub use global_message_handler::{GlobalActionHandler, GlobalMessage, GlobalMessageDiscriminant};
