+++ core/editor/src/input/mod.rs
@@ -1,0 +1,9 @@
+pub mod input_mapper;
+pub mod input_preprocessor;
+pub mod keyboard;
+pub mod mouse;
+
+pub use {
+	input_mapper::{InputMapper, InputMapperMessage, InputMapperMessageDiscriminant},
+	input_preprocessor::{InputPreprocessor, InputPreprocessorMessage, InputPreprocessorMessageDiscriminant},
+};
