+++ core/editor/src/input/mouse.rs
@@ -1,0 +1,48 @@
+use bitflags::bitflags;
+
+// origin is top left
+#[derive(Debug, Copy, Clone, Default, Eq, PartialEq)]
+pub struct ViewportPosition {
+	pub x: u32,
+	pub y: u32,
+}
+
+impl ViewportPosition {
+	pub fn distance(&self, other: &Self) -> f64 {
+		let x_diff = other.x as i32 - self.x as i32;
+		let y_diff = other.y as i32 - self.y as i32;
+		f64::sqrt((x_diff * x_diff + y_diff * y_diff) as f64)
+	}
+}
+
+#[derive(Debug, Copy, Clone, Default, Eq, PartialEq)]
+pub struct MouseState {
+	pub position: ViewportPosition,
+	pub mouse_keys: MouseKeys,
+}
+
+impl MouseState {
+	pub fn new() -> MouseState {
+		Self::default()
+	}
+
+	pub fn from_pos(x: u32, y: u32) -> MouseState {
+		MouseState {
+			position: ViewportPosition { x, y },
+			mouse_keys: MouseKeys::default(),
+		}
+	}
+	pub fn from_u8_pos(keys: u8, position: ViewportPosition) -> Self {
+		let mouse_keys = MouseKeys::from_bits(keys).expect("invalid modifier keys");
+		Self { position, mouse_keys }
+	}
+}
+bitflags! {
+	#[derive(Default)]
+	#[repr(transparent)]
+	pub struct MouseKeys: u8 {
+		const LEFT   = 0b0000_0001;
+		const RIGHT  = 0b0000_0010;
+		const MIDDLE = 0b0000_0100;
+	}
+}
