+++ core/editor/src/lib.rs
@@ -3,18 +3,17 @@
 
 extern crate graphite_proc_macros;
 
+mod communication;
 #[macro_use]
-mod macros;
-
-pub mod communication;
-pub mod derivable_custom_traits;
+mod misc;
 mod document;
-mod error;
-pub mod tools;
-pub mod workspace;
+mod frontend;
+mod global;
+pub mod input;
+pub mod tool;
 
 #[doc(inline)]
-pub use error::EditorError;
+pub use misc::EditorError;
 
 #[doc(inline)]
 pub use document_core::color::Color;
@@ -26,10 +25,7 @@ pub use document_core::LayerId;
 pub use document_core::document::Document as SvgDocument;
 
 #[doc(inline)]
-pub use communication::events;
-
-#[doc(inline)]
-pub use communication::Callback;
+pub use frontend::Callback;
 
 use communication::dispatcher::Dispatcher;
 // TODO: serialize with serde to save the current editor state
@@ -37,7 +33,7 @@ pub struct Editor {
 	dispatcher: Dispatcher,
 }
 
-use communication::message::prelude::*;
+use message_prelude::*;
 
 impl Editor {
 	pub fn new(callback: Callback) -> Self {
@@ -46,7 +42,21 @@ impl Editor {
 		}
 	}
 
-	pub fn handle_event(&mut self, event: events::Event) -> Result<(), EditorError> {
-		self.dispatcher.handle_message(InputPreprocessorMessage::Event(event).into())
+	pub fn handle_message<T: Into<Message>>(&mut self, message: T) -> Result<(), EditorError> {
+		self.dispatcher.handle_message(message)
 	}
+}
+
+pub mod message_prelude {
+	pub use super::communication::message::{AsMessage, Message, MessageDiscriminant};
+	pub use super::communication::MessageHandler;
+	pub use super::document::{DocumentMessage, DocumentMessageDiscriminant};
+	pub use super::frontend::{FrontendMessage, FrontendMessageDiscriminant};
+	pub use super::global::{GlobalMessage, GlobalMessageDiscriminant};
+	pub use super::input::{InputMapperMessage, InputMapperMessageDiscriminant, InputPreprocessorMessage, InputPreprocessorMessageDiscriminant};
+	pub use super::misc::derivable_custom_traits::{ToDiscriminant, TransitiveChild};
+	pub use super::tool::tool_messages::*;
+	pub use super::tool::tools::rectangle::{RectangleMessage, RectangleMessageDiscriminant};
+	pub use graphite_proc_macros::*;
+	pub use std::collections::VecDeque;
 }
