+++ core/editor/src/misc/mod.rs
@@ -1,0 +1,7 @@
+#[macro_use]
+pub mod macros;
+pub mod derivable_custom_traits;
+mod error;
+
+pub use error::EditorError;
+pub use macros::*;
