+++ core/editor/src/tool/tool_settings.rs
@@ -1,0 +1,20 @@
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum ToolSettings {
+	Select { append_mode: SelectAppendMode },
+	Ellipse,
+	Shape { shape: Shape },
+}
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum SelectAppendMode {
+	New,
+	Add,
+	Subtract,
+	Intersect,
+}
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+pub enum Shape {
+	Star { vertices: u32 },
+	Polygon { vertices: u32 },
+}
