+++ core/editor/src/tool/tools/mod.rs
@@ -1,0 +1,10 @@
+//mod crop;
+//mod ellipse;
+//mod eyedropper;
+//mod line;
+//mod navigate;
+//mod path;
+//mod pen;
+pub mod rectangle;
+//mod select;
+//mod shape;
