+++ core/editor/src/tool/tools/path.rs
