--- core/ptest/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "ptest"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-graphite-proc-macros = {path = "../proc-macro"}
